backend/requirements.txt
absl-py==0.9.0
appdirs==1.4.4
asgiref==3.2.10
astor==0.8.1
audioread==2.1.8
awscli==1.18.110
botocore==1.17.33
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
colorama==0.4.3
decorator==4.4.2
Django==3.0.8
django-cors-headers==3.4.0
django-rest-framework==0.1.0
djangorestframework==3.11.0
docutils==0.15.2
gast==0.2.2
google-auth==1.20.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.30.0
h5py==2.10.0
idna==2.10
importlib-metadata==1.7.0
jmespath==0.10.0
joblib==0.16.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
librosa==0.8.0
llvmlite==0.33.0
Markdown==3.2.2
networkx==2.2
numba==0.50.1
numpy==1.19.1
oauthlib==3.1.0
opt-einsum==3.3.0
packaging==20.4
pandas==1.0.3
pooch==1.1.1
protobuf==3.12.4
psycopg2-binary==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
resampy==0.2.2
rsa==4.5
s3transfer==0.3.3
scikit-learn==0.23.1
scipy==1.5.2
six==1.15.0
SoundFile==0.10.3.post1
SpeechRecognition==3.8.1
sqlparse==0.3.1
tensorboard==2.0.2
tensorflow==2.0.0
tensorflow-estimator==2.0.1
termcolor==1.1.0
threadpoolctl==2.1.0
urllib3==1.25.10
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0
